RCA Analysis study #DMAIC #SIXSIGMA #QA
Case:
Among the global clients that are part of the portfolio, we have one of great relevance, from the “brand and financial” point of view, which is going through a sensitive moment in its performance. This needs a quick diagnosis, because it has been suffering impacts not yet measured in its entirety.
The SAC service carried out by the Chat channel, until then, with 50 PA's was divided into two levels, with the first level a service for information with 30 PA's and the second level focused on resolution with 20 PA's, with greater authority and technical knowledge .
After this change, there was a drop in agent productivity in the consolidated result for the day, but looking at some times we still find moments of idleness.
When observing the results of the agents, there is a high dispersion (coefficient of variation 1.9) without any type of concentration, such as shift or leadership, but which have a strong correlation with BMT. As a consequence, the resolution and satisfaction results also suffered a decrease, with less dispersion, since the second level was given greater responsibility for solving problems.
It is important to emphasize that the historical results of the agents were considered to define the migrations, with the most experienced agents (longer time in attendance), faster and with a lower rate of monitoring errors being directed to level 2.
The client requested a quick action to identify the causes in search of the reversal of the three indicators simultaneously. As a priority, he asks that the employees' empathy be worked on, since the first level, when interacting with customers, does not adopt the best form of reception and/or responses are mechanical.
  In a logical and reasoned order, we describe a step-by-step process for quantified investigation (data), root cause analysis and action plan, considering the planning, human resources, quality and operational processes, not forgetting a sustainability plan . Studies for changing goals can be considered in your plan if they are substantiated.
CSAT target: 86% - Result: 66%
Goal Resolution: 80% - Result: 72%
Target productivity: 15 cases / hr – Result: 7 cases / hr